requirements.txt
streamlit
yfinance
scikit-learn
mysql-connector-python 
sqlalchemy 
pandas
python-dotenv
streamlit-login-auth-ui
trycourier
bcrypt 
streamlit-option-menu